Improve: Use Chrome's native lazy loading ability
This commit adds support for Chromium's native lazy loading ability (https://css-tricks.com/a-deep-dive-into-native-lazy-loading-for-images-and-frames/ or https://addyosmani.com/blog/lazy-loading/). 

It performs feature detection of support and if supported we set the src to all sources and images to their dataset equivalent (otherwise Chrome won't know what src to use!).

Else we fallback to lazysizes as before.